Perform file versioning with JSON file
Instead of saving e-tags to txt files, build a json file of URL metadata, like md5 hashes, e-tags, and paths where files are saved. Each subsequent time a file is requested, the e-tag in the json file is matched against the e-tag of the version currently downloadable from the URL. This method still ensures <250 characters in the absolute filepaths.