merge "Seafloorgrid contmask" branch (#103)
* Update SeafloorGrid continent masking

Allow pre-made continent masks to be passed. This skips the continent masking procedure. If masks are passed, the grid at min_time is used to make the initial ocean point mesh instead of a continent polygon file.

* Automate continent mask res downsizing

We use the continent mask at min_time instead of a continent polygon file when making the initial seafloor profile. We need to downsize the grid's resolution because this initial seafloor profile doesn't have to be fine - if it is, it slows down reconstruction by topologies

---------

Co-authored-by: lauren-ilano <lila0720@uni.sydney.edu.au>