Clip points slightly away from dateline (to avoid cartopy plot errors).
Because cartopy can sometimes move a point (of a dateline-wrapped
polygon or polyline) at central_meridian+180 to central_meridian-180
(or vice versa). This also means we no longer need to clean polygons
(introduced in dcc9499) - the pyGPlates dateline wrapper is sufficient
to wrap and tessellate polygons and polylines along the dateline.